requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
databases==0.8.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
boto3
pybranca